+++ py-endgame-aws/.devcontainer/devcontainer.json
@@ -56,5 +56,7 @@
     "source=${localEnv:HOME}${localEnv:USERPROFILE}/.aws/config,target=/root/.aws/config,type=bind,consistency=cached,ro=1",
     "source=${localEnv:HOME}${localEnv:USERPROFILE}/.claude,target=/root/.claude,type=bind,consistency=cached",
     "source=${localEnv:HOME}${localEnv:USERPROFILE}/.aws-batch/config.json,target=/root/.aws-batch/config.json,type=bind,consistency=cached,ro=1"
-  ]
+  ],
+  "postCreateCommand": "poetry install",
+  "postStartCommand": "endgame-aws -- --completion bash > ~/.endgame-aws-completion.sh"
 }
