+++ py-endgame-aws/main_test.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from main import plays
+from endgame_aws.cli import plays
 
 if __name__ == "__main__":
     asyncio.run(plays("womens", "2026-03-01"))
